requirement.txt
asgiref==3.8.1
Django==2.1.15
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
